97-colombian-online-shop/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
python-dotenv
stripe